enhance: code review
