+++ src/pages/Index.tsx
@@ -2,10 +2,10 @@
 import Navbar from "@/components/layout/Navbar";
 import Footer from "@/components/layout/Footer";
 import Hero from "@/components/sections/Hero";
-import Skills from "@/components/sections/skills/Skills";
 import Experience from "@/components/sections/Experience";
 import Projects from "@/components/sections/Projects";
 import Contact from "@/components/sections/Contact";
+import Skills from "@/components/sections/skills/Skills";
 
 const Index = () => {
   return (
